# doc/mall_admin/swagger.yaml
swagger: "2.0"
definitions:
...
  dto.DeleteDTO:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  dto.ResourceCreateDTO:
    properties:
      categoryId:
        description: 资源分类ID
        type: integer
      description:
        description: 描述
        type: string
      name:
        description: 资源名称
        type: string
      url:
        description: 资源URL
        type: string
    type: object
  dto.RoleAllocMenuDTO:
    properties:
      menuIds:
        description: 菜单id列表
        items:
          type: integer
        type: array
      roleId:
        description: 角色id
        type: integer
    type: object
  dto.RoleCreateDTO:
    properties:
      adminCount:
        description: 后台用户数量
        type: integer
      description:
        description: 角色描述
        type: string
      name:
        description: 角色名称
        type: string
      sort:
        description: 排序
        type: integer
      status:
        description: 启用状态：0->禁用；1->启用
        type: integer
    required:
    - name
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      message:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
...
  /resource/create:
    post:
      consumes:
      - application/json
      description: 创建资源
      parameters:
      - description: 创建资源
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.ResourceCreateDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 创建资源
      tags:
      - 资源
  /resource/delete/{id}:
    post:
      consumes:
      - application/json
      description: 删除资源
      parameters:
      - description: 资源ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 删除资源
      tags:
      - 资源
  /resource/list:
    get:
      consumes:
      - application/json
      description: 分页查询资源列表
      parameters:
      - description: 分页查询参数
        in: query
        name: data
        required: true
        type: dto.ResourcePageQueryDTO
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 分页查询资源列表
      tags:
      - 资源
  /resource/listAll:
    get:
      consumes:
      - application/json
      description: 查询所有资源列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 查询所有资源列表
      tags:
      - 资源
...
  /role/allocMenu:
    post:
      consumes:
      - application/json
      description: 给角色分配菜单
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 给角色分配菜单
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.RoleAllocMenuDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 给角色分配菜单
      tags:
      - 角色
  /role/allocResource:
    post:
      consumes:
      - application/json
      description: 给角色分配资源
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: integer
      - description: 给角色分配资源
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.RoleAllocMenuDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 给角色分配资源
      tags:
      - 角色
  /role/create:
    post:
      consumes:
      - application/json
      description: 创建角色
      parameters:
      - description: 创建角色
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.RoleCreateDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 创建角色
      tags:
      - 角色
  /role/delete:
    post:
      consumes:
      - application/json
      description: 删除角色
      parameters:
      - description: 角色ID
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteDTO'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 删除角色
      tags:
      - 角色
  /role/list:
    get:
      consumes:
      - application/json
      description: 分页查询角色列表
      parameters:
      - description: 分页查询参数
        in: query
        name: data
        required: true
        type: dto.PublicPageQuery
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 分页查询角色列表
      tags:
      - 角色
  /role/listAll:
    get:
      consumes:
      - application/json
      description: 查询所有角色列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
            type: object
        "400":
          description: 失败
          schema:
            $ref: '#/definitions/response.Response'
            type: object
      summary: 查询所有角色列表
      tags:
      - 角色
...